Move SettingsBuilder into another file
